+++ addon/utils/index.js
@@ -1,5 +1,11 @@
+/**
+ * Utility methods/modules
+ */
 import Ember from 'ember'
 const {$} = Ember
+
+export {default as events} from './events'
+export {default as keyCodes} from './key-codes'
 
 export function cloneEvent (event, target) {
   let newEvent = $.Event(null, event)
