+++ app/utils/key-codes.js
@@ -1,0 +1,1 @@
+export {keyCodes} from 'ember-frost-core/utils'
