--- app/utils/keycodes.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-frost-core/utils/keycodes'
